feat: Add user key pair generation and storage functionality
This commit adds the functionality to generate an asymmetric key pair and store it in IndexedDB. It introduces the `generateKeyPair` and `storeKeyPair` functions in the `app.js` file. The generated key pair is displayed in the browser using the `displayKeyInfo` function. The key pair can be generated by clicking the "Generate Key Pair" button and stored by clicking the "Store Key Pair in indexedDB" button.

This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes made.